feat(engine): P1-BE-001 modularize routes with blueprints
